V2.1 (#415)
Fixes for package generation + build/install/uninstall migration to scripts directory

+++ scripts/eosiocdt_uninstall.sh
@@ -14,6 +14,12 @@ binaries=(
    eosio-cc
    eosio-cpp
    eosio-ld
+   eosio-abidiff
+   eosio-init
+   llvm-readelf
+   llvm-objdump
+   llvm-ar
+   llvm-ranlib
 )
 
 if [ -d "/usr/local/eosio.cdt" ]; then
@@ -21,6 +27,10 @@ if [ -d "/usr/local/eosio.cdt" ]; then
    select yn in "Yes" "No"; do
       case $yn in
          [Yy]* )
+            if [ "$(id -u)" -ne 0 ]; then
+               printf "\nThis requires sudo, please run ./uninstall.sh with sudo\n\n"
+               exit -1
+            fi
             pushd /usr/local &> /dev/null
             rm -rf eosio.cdt
             pushd bin &> /dev/null
@@ -39,7 +49,7 @@ if [ -d "/usr/local/eosio.cdt" ]; then
    done
 fi
 
-if [ -d $OPT_LOCATION/eosio.cdt ]; then
+if [ -d "/usr/local/eosio.wasmsdk" ]; then
    printf "Do you wish to remove this install? (requires sudo)\n"
    select yn in "Yes" "No"; do
       case $yn in
@@ -48,6 +58,27 @@ if [ -d $OPT_LOCATION/eosio.cdt ]; then
                printf "\nThis requires sudo, please run ./uninstall.sh with sudo\n\n"
                exit -1
             fi
+            pushd /usr/local &> /dev/null
+            rm -rf eosio.wasmsdk
+            pushd bin &> /dev/null
+            for binary in ${binaries[@]}; do
+               rm ${binary}
+            done
+            popd &> /dev/null
+            break;;
+
+         [Nn]* ) 
+            printf "Aborting uninstall\n\n"
+            exit -1;;
+      esac
+   done
+fi
+
+if [ -d $OPT_LOCATION/eosio.cdt ] || || [ $1 == "force-new" ]; then
+   printf "Do you wish to remove this install?\n"
+   select yn in "Yes" "No"; do
+      case $yn in
+         [Yy]* )
             pushd $HOME &> /dev/null
             pushd opt &> /dev/null
             rm -rf eosio.cdt
@@ -63,27 +94,6 @@ if [ -d $OPT_LOCATION/eosio.cdt ]; then
             break;;
          [Nn]* )
             printf "\tAborting uninstall\n\n"
-            exit -1;;
-      esac
-   done
-fi
-
-if [ -d "/usr/local/eosio.wasmsdk" ]; then
-   printf "Do you wish to remove this install? (requires sudo)\n"
-   select yn in "Yes" "No"; do
-      case $yn in
-         [Yy]* )
-            pushd /usr/local &> /dev/null
-            rm -rf eosio.wasmsdk
-            pushd bin &> /dev/null
-            for binary in ${binaries[@]}; do
-               rm ${binary}
-            done
-            popd &> /dev/null
-            break;;
-
-         [Nn]* ) 
-            printf "Aborting uninstall\n\n"
             exit -1;;
       esac
    done
